+++ app/src/components/pages/__generated__/DeckQuery.ts
@@ -9,71 +9,7 @@ import { FlashCardStatus } from './../../../globalTypes'
 // GraphQL query operation: DeckQuery
 // ====================================================
 
-export interface DeckQuery_deck_notes_values_data_blocks_inlineStyleRanges {
-  __typename: 'InlineStyleRange'
-  style: string | null
-  offset: number | null
-  length: number | null
-}
-
-export interface DeckQuery_deck_notes_values_data_blocks_entityRanges {
-  __typename: 'EntityRange'
-  key: number
-  length: number
-  offset: number
-}
-
-export interface DeckQuery_deck_notes_values_data_blocks {
-  __typename: 'Block'
-  key: string
-  type: string
-  text: string
-  depth: number
-  inlineStyleRanges:
-    | (DeckQuery_deck_notes_values_data_blocks_inlineStyleRanges | null)[]
-    | null
-  entityRanges:
-    | (DeckQuery_deck_notes_values_data_blocks_entityRanges | null)[]
-    | null
-  data: any | null
-}
-
-export interface DeckQuery_deck_notes_values_data {
-  __typename: 'ContentState'
-  id: string
-  blocks: DeckQuery_deck_notes_values_data_blocks[]
-  entityMap: any
-}
-
-export interface DeckQuery_deck_notes_values_field {
-  __typename: 'Field'
-  /**
-   * Field id
-   */
-  id: string
-  /**
-   * Name of the field
-   */
-  name: string
-}
-
-export interface DeckQuery_deck_notes_values {
-  __typename: 'FieldValue'
-  /**
-   * Field value id
-   */
-  id: string
-  /**
-   * Field data
-   */
-  data: DeckQuery_deck_notes_values_data | null
-  /**
-   * Associated field
-   */
-  field: DeckQuery_deck_notes_values_field | null
-}
-
-export interface DeckQuery_deck_notes_model_primaryField {
+export interface DeckQuery_deck_notes_edges_node_model_primaryField {
   __typename: 'Field'
   /**
    * Field id
@@ -81,7 +17,7 @@ export interface DeckQuery_deck_notes_model_primaryField {
   id: string
 }
 
-export interface DeckQuery_deck_notes_model {
+export interface DeckQuery_deck_notes_edges_node_model {
   __typename: 'Model'
   /**
    * Name of this card model (e.g. "Basic", "Basic with Reversed")
@@ -91,10 +27,10 @@ export interface DeckQuery_deck_notes_model {
    * Primary field that should represent each individual note
    * of this model.
    */
-  primaryField: DeckQuery_deck_notes_model_primaryField | null
+  primaryField: DeckQuery_deck_notes_edges_node_model_primaryField | null
 }
 
-export interface DeckQuery_deck_notes_flashCards_template {
+export interface DeckQuery_deck_notes_edges_node_flashCards_template {
   __typename: 'Template'
   /**
    * Name of the template
@@ -102,7 +38,7 @@ export interface DeckQuery_deck_notes_flashCards_template {
   name: string | null
 }
 
-export interface DeckQuery_deck_notes_flashCards {
+export interface DeckQuery_deck_notes_edges_node_flashCards {
   __typename: 'FlashCard'
   /**
    * FlashCard id.
@@ -125,10 +61,10 @@ export interface DeckQuery_deck_notes_flashCards {
   /**
    * Template associated with this flashcard.
    */
-  template: DeckQuery_deck_notes_flashCards_template | null
+  template: DeckQuery_deck_notes_edges_node_flashCards_template | null
 }
 
-export interface DeckQuery_deck_notes_deck {
+export interface DeckQuery_deck_notes_edges_node_deck {
   __typename: 'Deck'
   /**
    * Title of the deck
@@ -136,28 +72,85 @@ export interface DeckQuery_deck_notes_deck {
   title: string
 }
 
-export interface DeckQuery_deck_notes {
+export interface DeckQuery_deck_notes_edges_node {
   __typename: 'Note'
   /**
    * Note id
    */
   id: string
   /**
-   * Values of this note
+   * Note text representation
    */
-  values: DeckQuery_deck_notes_values[]
+  text: string
   /**
    * Model of this note
    */
-  model: DeckQuery_deck_notes_model | null
+  model: DeckQuery_deck_notes_edges_node_model | null
   /**
    * Generated flashcards
    */
-  flashCards: DeckQuery_deck_notes_flashCards[]
+  flashCards: DeckQuery_deck_notes_edges_node_flashCards[]
   /**
    * Deck containing this note
    */
-  deck: DeckQuery_deck_notes_deck | null
+  deck: DeckQuery_deck_notes_edges_node_deck | null
+}
+
+export interface DeckQuery_deck_notes_edges {
+  __typename: 'NoteEdge'
+  node: DeckQuery_deck_notes_edges_node
+  cursor: string
+}
+
+export interface DeckQuery_deck_notes_pageInfo {
+  __typename: 'PageInfo'
+  hasNextPage: boolean
+  hasPreviousPage: boolean
+  endCursor: string | null
+  startCursor: string | null
+}
+
+export interface DeckQuery_deck_notes_pageCursors_first {
+  __typename: 'PageCursor'
+  cursor: string
+  page: number
+  isCurrent: boolean
+}
+
+export interface DeckQuery_deck_notes_pageCursors_around {
+  __typename: 'PageCursor'
+  cursor: string
+  page: number
+  isCurrent: boolean
+}
+
+export interface DeckQuery_deck_notes_pageCursors_last {
+  __typename: 'PageCursor'
+  cursor: string
+  page: number
+  isCurrent: boolean
+}
+
+export interface DeckQuery_deck_notes_pageCursors_previous {
+  __typename: 'PageCursor'
+  cursor: string
+  page: number
+  isCurrent: boolean
+}
+
+export interface DeckQuery_deck_notes_pageCursors {
+  __typename: 'PageCursors'
+  first: DeckQuery_deck_notes_pageCursors_first | null
+  around: DeckQuery_deck_notes_pageCursors_around[]
+  last: DeckQuery_deck_notes_pageCursors_last | null
+  previous: DeckQuery_deck_notes_pageCursors_previous | null
+}
+
+export interface DeckQuery_deck_notes {
+  __typename: 'NoteConnection'
+  edges: DeckQuery_deck_notes_edges[]
+  pageInfo: DeckQuery_deck_notes_pageInfo
+  pageCursors: DeckQuery_deck_notes_pageCursors
 }
 
 export interface DeckQuery_deck {
@@ -189,7 +182,7 @@ export interface DeckQuery_deck {
   /**
    * Notes contained in this deck
    */
-  notes: (DeckQuery_deck_notes | null)[] | null
+  notes: DeckQuery_deck_notes | null
 }
 
 export interface DeckQuery {
@@ -201,4 +194,6 @@ export interface DeckQuery {
 
 export interface DeckQueryVariables {
   slug: string
+  page: number
+  size: number
 }