+++ app/src/hooks/useLatestRefEffect.ts
@@ -1,0 +1,23 @@
+import { useEffect, useRef } from 'react'
+
+function useLatestRefEffect<T>(value: T, callback: (value: T) => void) {
+  const latestRef = useRef(value)
+
+  const callbackRef = useRef(callback)
+
+  useEffect(() => {
+    callbackRef.current = callback
+  }, [callback])
+
+  useEffect(() => {
+    if (latestRef.current === value) {
+      return
+    }
+
+    latestRef.current = value
+
+    callbackRef.current(value)
+  }, [value])
+}
+
+export default useLatestRefEffect